Basic cleanup: fix CI, and test with 9.10.1
There was some conflict between the bionic image that we were using and
the environment in which the checkout action runs, so we were not able
to even check the code out in Github action. Rather embarrassing.

We've been supporting GHC 9.10, but didn't test, so it was a good
opportunity to add it to the matrix.